[skip ci] add support for a dedicated builder (#1)
Travis is too slow and does not work reliably for the builds.

+++ scripts/patch.sh
@@ -14,31 +14,39 @@ LSTFILE=versions.txt
 mkdir -p tarballs
 
 for tarball in $(grep '# tarball' $KOLLA_CONF_FILE | awk '{ print $4 }'); do
-    pushd tarballs
-    wget --quiet $tarball
+    pushd tarballs > /dev/null
+
     filename=$(basename $tarball)
+    if [[ -e $filename ]]; then
+        popd > /dev/null
+        continue
+    fi
+
+    echo Download $tarball
+    wget --quiet $tarball
 
     if [[ $tarball == *"gnocchi"* && ! $filename == *"gnocchi"* ]]; then
         mv $filename gnocchi-$filename
         filename="gnocchi-$filename"
     fi
 
+    echo Process $filename
     tar xzf $filename
     directory=$(tar -tzf $filename | head -1 | cut -f1 -d"/")
     echo $directory >> $LSTFILE
     rm $filename
     if [[ -e ../patches/$OPENSTACK_VERSION/${directory%-*} ]]; then
-        pushd $directory
+        pushd $directory > /dev/null
         for patch in $(find ../../patches/$OPENSTACK_VERSION/${directory%-*} -type f -name '*.patch'); do
             echo "APPLY PATCH $patch"
             patch --forward --batch -p1 --dry-run < $patch || exit 1
             patch --forward --batch -p1 < $patch
         done
-        popd
+        popd > /dev/null
     fi
     tar czf $filename $directory
     rm -r $directory
-    popd
+    popd > /dev/null
 done
 
 cat tarballs/$LSTFILE | sort | uniq > $LSTFILE
